refactor: update letter spacing in LoginComponent and enhance NavBarProfile structure
