+++ app/components/CardComponent/SecondaryCard.tsx
@@ -1,4 +1,4 @@
-// import Link from "next/link";
+"use client";
 import React, { FC, ReactNode } from "react";
 import MoreVertIcon from "@mui/icons-material/MoreVert";
 
@@ -8,6 +8,7 @@ interface SecondaryCardProps {
   footer?: ReactNode;
   children?: ReactNode;
   className?: string;
+  onClick?: () => void;
 }
 
 const SecondaryCard: FC<SecondaryCardProps> = ({
@@ -16,36 +17,39 @@ const SecondaryCard: FC<SecondaryCardProps> = ({
   footer,
   children,
   className = "",
+  onClick,
 }) => {
   return (
     <div className={className}>
-      {header ? (
-        <header
-          className="relative bg-cover bg-center p-5 rounded-t-2xl"
-          style={{ backgroundImage: "url(/card_bg.jpeg)" }}
-        >
-          <div className="absolute top-6 right-4 text-white cursor-pointer">
-            <MoreVertIcon />
+      <div onClick={onClick} className="cursor-pointer">
+        {header ? (
+          <header
+            className="relative bg-cover bg-center p-5 rounded-t-2xl"
+            style={{ backgroundImage: "url(/card_bg.jpeg)" }}
+          >
+            <div className="absolute top-6 right-4 text-white cursor-pointer">
+              <MoreVertIcon />
+            </div>
+            {header}
+          </header>
+        ) : (
+          <div className="relative p-5 bg-gray-200 rounded-t-2xl">
+            <div className="absolute top-4 right-4 text-white cursor-pointer">
+              <MoreVertIcon />
+            </div>
+            <h2 className="text-xl font-semibold text-gray-800 mb-2">
+              Default Header
+            </h2>
           </div>
-          {header}
-        </header>
-      ) : (
-        <div className="relative p-5 bg-gray-200 rounded-t-2xl">
-          <div className="absolute top-4 right-4 text-white cursor-pointer">
-            <MoreVertIcon />
-          </div>
-          <h2 className="text-xl font-semibold text-gray-800 mb-2">
-            Default Header
-          </h2>
-        </div>
-      )}
-      {content ? (
-        <main className="flex flex-col gap-2 px-5">{content}</main>
-      ) : (
-        <p className="text-gray-600 mb-4">
-          This is the default content of the card.
-        </p>
-      )}
+        )}
+        {content ? (
+          <main className="flex flex-col gap-2 px-5">{content}</main>
+        ) : (
+          <p className="text-gray-600 mb-4">
+            This is the default content of the card.
+          </p>
+        )}
+      </div>
 
       {footer ? (
         <footer className="border-t-2 px-5 py-2 text-center flex items-center justify-between">
